New Commands: 'clean' and 'clean_all'
New commands that will allow tfgent to delete files that were generated
by a template. These commands expect a 'clean_pattern' to be defined,
ideally in the root configuration to allow for the 'clean_all' command,
which will clean the 'child_directories' defined in the root confing.
This should give the user a great deal of control around how they want
things cleaned up when needed.